requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.20.0
PyYAML>=6.0

# Data processing
Pillow>=8.0.0
nltk>=3.7
numpy>=1.21.0

# Evaluation metrics
gensim>=4.0.0
scikit-learn>=1.0.0

# API and utilities
openai>=0.27.0
tqdm>=4.60.0

# Optional: For better performance
accelerate>=0.20.0
